// src/app/models/animal.ts
export interface Animal {
  id: String,
  description: String;
  mainPicture: String;
  secondPicture: String;
  thirdPicture: String;
  fourthPicture: String;
  category: String;
  specie: String;
  age: number;
  price: number;
  sold: boolean;
};
export interface Article {
  id: String,
  specie: String;
  age: number;
  price: number;
};